+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [1.1.0] - 2019-03-08
+
+### Changed
+- es6 support 
+- make usage of `verlok/lazyload` version 11
+- drop element `lazyload:set` event (use `lazyload:reveal` event instead)
+- add `window.lazyImagesInstance` with `update`, `updateImages` and `updateWrappers` function for usage in 3rd party plugins
+
 ## [1.0.1] - 2019-03-07
 
 ### Fixed